+++ docs/changes.rst
@@ -5,6 +5,8 @@ Next version
 ------------
 
 * Removed support for Django < 3.2.
+* Updated check ``W006`` to look for
+  ``django.template.loaders.app_directories.Loader``.
 
 3.2.4 (2021-12-15)
 ------------------